# Speed up zipWith some more